feature-policy/reporting/screen-wake-lock-report-only.https.html

<!DOCTYPE html>
<html>
  <head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
  </head>
  <body>
    <script>
      promise_test(async (t) => {
        const reportPromise = new Promise(resolve => {
          new ReportingObserver((reports, observer) => resolve(reports),
                                {types: ["feature-policy-violation"]}).observe();
        });
        await navigator.wakeLock.request('screen');
        const reports = await reportPromise;

        assert_equals(reports.length, 1);
        const report = reports[0];
        assert_equals(report.type, "feature-policy-violation");
        assert_equals(report.body.featureId, "screen-wake-lock");
        assert_equals(report.body.disposition, "report");
      }, "Screen Wake Lock report only mode");
    </script>
  </body>
</html>
